Update README.md
Solar Lamp experiments

I'm using the popular YX8089 DC DC Converter / Joule Thief, 
the way like to think about this device is that it can get the last bits of energy out of 
a battery or solar cell.
Normally the battery is in parallel with the solar cell, and this configuration pulls the solar cell voltage low until the battery is charged.
I wondered if having the joule thief
charge a battery might be better. So the solar cell can generate a full o/p voltage.
So have the battery after the joule thief.
Then the problem is that the battery will not drive the LED, so we'll need another Joule thief, one to take power from the
solar cell, to feed in to the Battery, and another one to take energy from the battery to drive the LED.
Then I was wondering, maybe it might be of use to have a 7805 V Reg output from the 2nd battery joule thief combination,
or maybe some other regultor.
So what I'v tried to do is create a circuit that could serve as a single joule thief circuit as in a regular solar lamp,
but build on to the board pads so that another joule thief can be fitted, along with a regulated output.
Purely experimental at the moment. 
One problem is the output of the YX8089 can be as high as 100V as a result of flyback voltages from the inductor, which will obviously
either over charge any battery, or damage any regulator.
Ordinarily in an inductor circuit there is a diode to limit flyback voltages, so I'm wondering if a Zener could be substituted here,
Let's say a 15 Volt Zener, would this limit the output pulses to 15V, instead of 100V.
That might protect the Volt Reg.
It might turn out that nothing useful can come from using two joule thieves in series and a regulator, that being the case,
I'll still have  a regular Solar lamp circuit which I can hook up to a bigger solar cell than they're normally fitted with.